命中特效、特效urp設定

+++ Assets/Script/StateMachine/Enemy/EnemyAttack.cs
@@ -10,5 +10,5 @@ public class EnemyAttack
     [field: SerializeField, HeaderAttribute("動畫過渡時間")] public float TransitionDuration { get; private set; } // 過渡時間
     [field: SerializeField, HeaderAttribute("傷害值")] public int Damage { get; private set; } // 傷害值
     [field: SerializeField, HeaderAttribute("San傷害值")] public int SanDamage { get; private set; } // San傷害值
-    [field: SerializeField, HeaderAttribute("硬直時間")] public int CooldownTime { get; private set; }
+    [field: SerializeField, HeaderAttribute("硬直時間")] public float CooldownTime { get; private set; }
 }
